+++ packages/nextclaw-remote/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nextclaw/remote
 
+## 0.1.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @nextclaw/server@0.10.55
+
 ## 0.1.48
 
 ### Patch Changes